requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
jinja2==3.1.6
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.4
pydantic-core==2.41.5
